updated App to use index files
