+++ src/components/SubmittedRequestsTable/Sort.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Sort = () => {
+  return <div>test</div>;
+};
