+++ src/components/index.tsx
@@ -2,4 +2,4 @@ export { GetMinistryAcronyms } from "./API/GET/GetMinistryAcronyms";
 export { GetSubmittedRequests } from "./API/GET/GetSubmittedRequests";
 
 // Components
-export { SubmittedRequestsTable } from "components/SubmittedRequestsTable/index";
+export * from "components/SubmittedRequestsTable";
